Refactor url of page and move component to components folder
